+++ src/main/java/com/teletubbies/course/course/CourseCategoryTypeConverter.java
@@ -6,15 +6,15 @@ import java.util.Optional;
 
 @Converter(autoApply = true)
 public class CourseCategoryTypeConverter
-    implements AttributeConverter<CourseCategoryType, Integer> {
+    implements AttributeConverter<CourseCategoryType, Short> {
 
   @Override
-  public Integer convertToDatabaseColumn(final CourseCategoryType attribute) {
+  public Short convertToDatabaseColumn(final CourseCategoryType attribute) {
     return Optional.ofNullable(attribute).map(CourseCategoryType::getKey).orElse(null);
   }
 
   @Override
-  public CourseCategoryType convertToEntityAttribute(final Integer dbData) {
+  public CourseCategoryType convertToEntityAttribute(final Short dbData) {
     return Optional.ofNullable(dbData).map(CourseCategoryType::fromKey).orElse(null);
   }
 }
