+++ src/main/java/com/teletubbies/course/course/CourseLevelTypeConverter.java
@@ -5,15 +5,15 @@ import jakarta.persistence.Converter;
 import java.util.Optional;
 
 @Converter(autoApply = true)
-public class CourseLevelTypeConverter implements AttributeConverter<CourseLevelType, Integer> {
+public class CourseLevelTypeConverter implements AttributeConverter<CourseLevelType, Short> {
 
   @Override
-  public Integer convertToDatabaseColumn(final CourseLevelType attribute) {
+  public Short convertToDatabaseColumn(final CourseLevelType attribute) {
     return Optional.ofNullable(attribute).map(CourseLevelType::getKey).orElse(null);
   }
 
   @Override
-  public CourseLevelType convertToEntityAttribute(final Integer dbData) {
+  public CourseLevelType convertToEntityAttribute(final Short dbData) {
     return Optional.ofNullable(dbData).map(CourseLevelType::fromKey).orElse(null);
   }
 }
